make sure migration 1 doesn't remove other page keys
